Implement SOLID principle by collaboration
